backend/pyproject.toml
[tool.poetry]
name = "kiros-triage-be"
version = "0.1.0"
description = ""
authors = ["Lê Ngọ <ngole719@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.0"}
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.40"
alembic = "^1.15.0"
psycopg2-binary = "^2.9.10"
redis = "^5.2.0"
pydantic-settings = "^2.8.0"
typer = "^0.15.0"
python-dotenv = "^1.0.0"
sqlalchemy-utils = "^0.42.0"
google-genai = "^1.61.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"